Minor changes renamed Icon-Settings.tsx to Icon_Settings.tsx fixed settings icon, clip-path to clipPath removed unneeded url prop form ConnectionContext and main.tsx
