+++ electron-vite-project/src/assets/Icon_Settings.tsx
@@ -17,7 +17,7 @@ const IconSettings: React.FC<IconSettingsProps> = ({ color = 'currentColor', siz
         fill='none'
         xmlns='http://www.w3.org/2000/svg'
       >
-        <g clip-path='url(#clip0_2207_34745)'>
+        <g clipPath='url(#clip0_2207_34745)'>
           <path
             d='M12.3618 26.1289H14.729C15.6313 26.1289 16.3227 25.5781 16.5336 24.7109L17.0375 22.5195L17.4125 22.3906L19.3227 23.5625C20.0844 24.043 20.9633 23.9258 21.6079 23.2812L23.2485 21.6523C23.893 21.0078 24.0102 20.1172 23.5297 19.3672L22.3344 17.4687L22.4751 17.1172L24.6665 16.6015C25.5219 16.3906 26.0844 15.6875 26.0844 14.7969V12.4765C26.0844 11.5859 25.5336 10.8828 24.6665 10.6719L22.4985 10.1445L22.3461 9.76953L23.5415 7.8711C24.0219 7.1211 23.9047 6.24219 23.2602 5.58594L21.6196 3.94531C20.9868 3.3125 20.1079 3.19531 19.3461 3.66406L17.436 4.83594L17.0375 4.6836L16.5336 2.49219C16.3227 1.625 15.6313 1.07422 14.729 1.07422H12.3618C11.4594 1.07422 10.7681 1.625 10.5571 2.49219L10.0415 4.6836L9.64307 4.83594L7.74463 3.66406C6.98291 3.19531 6.09229 3.3125 5.45947 3.94531L3.83057 5.58594C3.18604 6.24219 3.05713 7.1211 3.54932 7.8711L4.73291 9.76953L4.59229 10.1445L2.42432 10.6719C1.55713 10.8828 1.00635 11.5859 1.00635 12.4765V14.7969C1.00635 15.6875 1.56885 16.3906 2.42432 16.6015L4.61573 17.1172L4.74463 17.4687L3.56104 19.3672C3.06885 20.1172 3.19776 21.0078 3.84229 21.6523L5.47119 23.2812C6.11573 23.9258 7.00635 24.043 7.76807 23.5625L9.66651 22.3906L10.0415 22.5195L10.5571 24.7109C10.7681 25.5781 11.4594 26.1289 12.3618 26.1289ZM12.5493 24.3008C12.3501 24.3008 12.2446 24.2187 12.2094 24.0312L11.5063 21.125C10.7915 20.9492 10.1235 20.668 9.61963 20.3515L7.06494 21.9219C6.92432 22.0273 6.77197 22.0039 6.63135 21.8633L5.24854 20.4804C5.11963 20.3515 5.10791 20.2109 5.20166 20.0469L6.77197 17.5156C6.50244 17.0234 6.19776 16.3554 6.01026 15.6406L3.10401 14.9492C2.91651 14.914 2.83447 14.8086 2.83447 14.6094V12.6523C2.83447 12.4414 2.90479 12.3476 3.10401 12.3125L5.99854 11.6094C6.18604 10.8477 6.5376 10.1562 6.74854 9.72266L5.18994 7.19141C5.08447 7.01563 5.09619 6.875 5.2251 6.73438L6.61963 5.375C6.76026 5.23438 6.88916 5.21094 7.06494 5.31641L9.59619 6.85156C10.1001 6.57031 10.8149 6.27734 11.518 6.07813L12.2094 3.17188C12.2446 2.98438 12.3501 2.90234 12.5493 2.90234H14.5415C14.7407 2.90234 14.8461 2.98438 14.8696 3.17188L15.5844 6.10156C16.311 6.28906 16.9438 6.58203 17.4711 6.86328L20.0141 5.31641C20.2016 5.21094 20.3188 5.23438 20.4711 5.375L21.854 6.73438C21.9946 6.875 21.9946 7.01563 21.8891 7.19141L20.3305 9.72266C20.5532 10.1562 20.893 10.8477 21.0805 11.6094L23.9868 12.3125C24.1743 12.3476 24.2563 12.4414 24.2563 12.6523V14.6094C24.2563 14.8086 24.1625 14.914 23.9868 14.9492L21.0688 15.6406C20.8813 16.3554 20.5883 17.0234 20.3071 17.5156L21.8774 20.0469C21.9711 20.2109 21.9711 20.3515 21.8305 20.4804L20.4594 21.8633C20.3071 22.0039 20.1665 22.0273 20.0141 21.9219L17.4594 20.3515C16.9555 20.668 16.2993 20.9492 15.5844 21.125L14.8696 24.0312C14.8461 24.2187 14.7407 24.3008 14.5415 24.3008H12.5493ZM13.5454 18.0781C15.9946 18.0781 18.0102 16.0625 18.0102 13.6015C18.0102 11.164 15.9946 9.14844 13.5454 9.14844C11.0961 9.14844 9.06885 11.164 9.06885 13.6015C9.06885 16.0508 11.0844 18.0781 13.5454 18.0781ZM13.5454 16.2617C12.0922 16.2617 10.897 15.0664 10.897 13.6015C10.897 12.1601 12.0922 10.9648 13.5454 10.9648C14.9751 10.9648 16.1704 12.1601 16.1704 13.6015C16.1704 15.0547 14.9751 16.2617 13.5454 16.2617Z'
             fill={color}
