+++ electron-vite-project/src/components/Settings.tsx
@@ -22,7 +22,7 @@ import { useConnectionContext } from './ConnectionContext';
 import AdminPassword from './AdminPassword';
 import Setting1 from './Settings_General';
 import Setting2 from './Settings_Admin';
-import IconSettings from '@/assets/Icon-Settings';
+import IconSettings from '@/assets/Icon_Settings';
 import IconAdmin from '@/assets/Icon_Admin';
 import './Titlebar.css';
